#117 Minor cleanup
