+++ core/src/main/java/contacts/core/entities/Entity.kt
@@ -1,7 +1,6 @@
 package contacts.core.entities
 
 import android.os.Parcelable
-import java.util.Comparator
 
 /**
  * Base of all entities provided in this library.
@@ -13,8 +12,15 @@ sealed interface Entity : Parcelable {
      */
     val isBlank: Boolean
 
+    /**
+     * Only existing data entities have an id. For all others, this will return null.
+     */
+    // This is declared here instead of an extension function for easier use for Java users.
+    val idOrNull: Long?
+        get() = null
+
     companion object {
-        const val INVALID_ID: Long = -1L
+        internal const val INVALID_ID: Long = -1L
     }
 }
 
@@ -39,6 +45,9 @@ sealed interface ExistingEntity : Entity {
      * The ID of this entity (row) in the table it belongs to.
      */
     val id: Long
+
+    override val idOrNull: Long?
+        get() = id
 }
 
 /**
@@ -156,42 +165,4 @@ internal fun entitiesAreAllBlank(vararg collectionOfEntities: Collection<Entity>
         }
     }
     return true
-}
-
-/**
- * Only existing data entities have an id. For all others, this will return null.
- */
-val Entity.idOrNull: Long?
-    get() = when (this) {
-        is ExistingEntity -> id
-        else -> null
-    }
-
-/**
- * Entities can either be new or existing entities. Only existing entities have IDs. Therefore, when
- * comparing by ID, we have to be able to compare null and non-null ids.
- */
-// Expose this to consumers. Maybe they'll find a use for it?
-object EntityIdComparator : Comparator<Entity> {
-
-    override fun compare(o1: Entity, o2: Entity): Int = o1.idOrNull.compareTo(o2.idOrNull)
-
-    /**
-     * Compares [this] nullable long to the [other] nullable long.
-     *
-     * If both [this] and [other] are not null, then a comparison is done on both. Otherwise, this
-     * returns a positive integer if [this] is null and [other] is not null. Returns a negative integer
-     * if [this] is not null and [other] is null. Returns 0 if both [this] and [other] are null.
-     */
-    private fun Long?.compareTo(other: Long?): Int {
-        return if (this != null && other != null) {
-            compareTo(other)
-        } else if (this == null && other != null) {
-            1
-        } else if (this != null && other == null) {
-            -1
-        } else {
-            0
-        }
-    }
 }
