+++ core/src/main/java/contacts/core/util/ContactRefresh.kt
@@ -18,8 +18,7 @@ import contacts.core.equalTo
  *
  * ## Permissions
  *
- * The [contacts.core.ContactsPermissions.READ_PERMISSION] is required. Otherwise, null will be returned
- * if the permission is not granted.
+ * The [contacts.core.ContactsPermissions.READ_PERMISSION] is required.
  *
  * ## Thread Safety
  *
