+++ core/src/main/java/contacts/core/util/DataRefresh.kt
@@ -14,8 +14,7 @@ import contacts.core.entities.*
  *
  * ## Permissions
  *
- * The [contacts.core.ContactsPermissions.READ_PERMISSION] is required. Otherwise, null will be
- * returned if the permission is not granted.
+ * The [contacts.core.ContactsPermissions.READ_PERMISSION] is required.
  *
  * ## Thread Safety
  *
